pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dev-notes-cli"
version = "0.1.0"
description = "A lightweight CLI tool for generating structured Markdown notes for technical blogs and study logs."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "wenzy" }
]
dependencies = [
  "typer>=0.12.0",
  "PyYAML>=6.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0"
]

[project.scripts]
devnotes = "devnotes.cli:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.setuptools]
packages = ["devnotes"]
